add abstraction for database connection
